# perf(utils/disk-cache): add a fs-persistent cache