Added in `DiskCache` which persists caching across app restarts, which is useful with expensive operations. It'll allow us to restart server often without losing the cached data. It's a database, but without the hassle of a database.

This filesystem-backed cache is used internally in [respec-github-apis](https://github.com/sidvishnoi/respec-github-apis) as [`TTLCache`](https://github.com/sidvishnoi/respec-github-apis/blob/main/utils/cache.ts). I'm bringing it here in this PR as a step to move respec-github-apis into this repo, similar to https://github.com/marcoscaceres/respec.org/pull/146 and https://github.com/marcoscaceres/respec.org/pull/148.

As a demonstration, I converted `/github/org/repo/contributors` endpoint to use the disk cache:

|                    | Before                              | After                               |
| ------------------ | ----------------------------------- | ----------------------------------- |
| **First app load** |                                     |                                     |
|                    | `2573.480 ms # cache miss`          | `2452.966 ms # cache miss`          |
|                    | ` 1.152 ms # cached in memory`      | ` 1.179 ms # cached in memory`      |
| **Restart app**    |                                     |                                     |
|                    | `2538.911 ms # cache miss`          | ` 6.387 ms # cache hit from disk`   |
|                    | ` 1.277 ms # cache hit from memory` | ` 1.056 ms # cache hit from memory` |